requirements.txt
flask>=3.0.0
streamlit>=1.34.0
pandas>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.1
stable-baselines3>=2.3.0
pyserial>=3.5
sumolib>=1.20.0
traci>=1.20.0
